pyproject.toml
[tool.poetry]
name = "drf-web3-auth"
version = "0.1.0"
description = "DRF app with web3 authorization"
authors = ["Vasiliy Ivanov <shitcoder@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
django-rest-framework = "^0.1.0"
web3 = "^6.12.0"
celery = "^5.3.6"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
django-extensions = "^3.2.3"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
eth-account = "^0.10.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}
pytest = "^7.4.3"


[tool.poetry.group.dev.dependencies]
ipython = "^8.18.1"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
